Don't recreate MapboxInspect control when switching popup on hover
When switching the state of "popup on hover", instead of recreating MapboxInspect control, we only change the boolean in the control's option and re-render. This fixes an issue of not being to switching the inspect style after switching the "popup on hover".

Additionally, use a radio group for the here map style switch instead of using two buttons.